Complete refactor to allow web support
This large commit refactors and changes the implementation of quite a bit in order to make there only be one running loop and only one thread. This lets this game be deployable in the browser with the use of Pygbag. There's still a lot of work to be done, both in fixing bugs and improving UX/performance.